Have a skinny controller main
