--- cmd/controller/signal_windows.go
@@ -1,3 +1,0 @@
-package main
-
-func initKeyGenSignalListener(trigger func()) {}
