+++ cmd/kubeseal/main.go
@@ -50,31 +50,31 @@ func bindFlags(f *kubeseal.Flags, fs *flag.FlagSet) {
 	fs.Var(&f.SealingScope, "scope", "Set the scope of the sealed secret: strict, namespace-wide, cluster-wide (defaults to strict). Mandatory for --raw, otherwise the 'sealedsecrets.bitnami.com/cluster-wide' and 'sealedsecrets.bitnami.com/namespace-wide' annotations on the input secret can be used to select the scope.")
 	fs.BoolVar(&f.ReEncrypt, "rotate", false, "")
 	fs.BoolVar(&f.ReEncrypt, "re-encrypt", false, "Re-encrypt the given sealed secret to use the latest cluster key.")
-	_ = flag.CommandLine.MarkDeprecated("rotate", "please use --re-encrypt instead")
+	_ = fs.MarkDeprecated("rotate", "please use --re-encrypt instead")
 
 	fs.BoolVar(&f.Unseal, "recovery-unseal", false, "Decrypt a sealed secrets file obtained from stdin, using the private key passed with --recovery-private-key. Intended to be used in disaster recovery mode.")
 	fs.StringSliceVar(&f.PrivKeys, "recovery-private-key", nil, "Private key filename used by the --recovery-unseal command. Multiple files accepted either via comma separated list or by repetition of the flag. Either PEM encoded private keys or a backup of a json/yaml encoded k8s sealed-secret controller secret (and v1.List) are accepted. ")
 }
 
-func bindClientFlags(overrides *clientcmd.ConfigOverrides) {
-	flagenv.SetFlagsFromEnv(flagEnvPrefix, goflag.CommandLine)
+func bindClientFlags(fs *flag.FlagSet, gofs *goflag.FlagSet, overrides *clientcmd.ConfigOverrides) {
+	flagenv.SetFlagsFromEnv(flagEnvPrefix, gofs)
 
-	initUsualKubectlFlags(overrides, flag.CommandLine)
+	initUsualKubectlFlags(overrides, fs)
 
-	pflagenv.SetFlagsFromEnv(flagEnvPrefix, flag.CommandLine)
+	pflagenv.SetFlagsFromEnv(flagEnvPrefix, fs)
 
 	// add klog flags to goflags flagset
 	klog.InitFlags(nil)
 	// Standard goflags (glog in particular)
-	flag.CommandLine.AddGoFlagSet(goflag.CommandLine)
+	fs.AddGoFlagSet(gofs)
 }
 
-func initUsualKubectlFlags(overrides *clientcmd.ConfigOverrides, flagset *flag.FlagSet) {
+func initUsualKubectlFlags(overrides *clientcmd.ConfigOverrides, fs *flag.FlagSet) {
 	kflags := clientcmd.RecommendedConfigOverrideFlags("")
-	clientcmd.BindOverrideFlags(overrides, flagset, kflags)
+	clientcmd.BindOverrideFlags(overrides, fs, kflags)
 }
 
-func mainE(w io.Writer, fs *flag.FlagSet, args []string) error {
+func mainE(w io.Writer, fs *flag.FlagSet, gofs *goflag.FlagSet, args []string) error {
 	var flags kubeseal.Flags
 	var printVersion bool
 	var overrides clientcmd.ConfigOverrides
@@ -82,11 +82,11 @@ func mainE(w io.Writer, fs *flag.FlagSet, args []string) error {
 
 	fs.BoolVar(&printVersion, "version", false, "Print version information and exit")
 	bindFlags(&flags, fs)
-	bindClientFlags(&overrides)
+	bindClientFlags(fs, gofs, &overrides)
 	if err := fs.Parse(args); err != nil {
 		return err
 	}
-	if err := goflag.CommandLine.Parse([]string{}); err != nil {
+	if err := gofs.Parse([]string{}); err != nil {
 		return err
 	}
 
@@ -101,7 +101,7 @@ func mainE(w io.Writer, fs *flag.FlagSet, args []string) error {
 }
 
 func main() {
-	if err := mainE(os.Stdout, flag.CommandLine, os.Args); err != nil {
+	if err := mainE(os.Stdout, flag.CommandLine, goflag.CommandLine, os.Args); err != nil {
 		fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
 	}
