+++ pkg/controller/signal_windows.go
@@ -1,0 +1,3 @@
+package controller
+
+func initKeyGenSignalListener(trigger func()) {}
